backend/requirements.txt
# Core
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3

# Database
psycopg2-binary>=2.9

# Security
django-axes>=6.4

# Config management
python-dotenv>=1.0

# Production WSGI server
gunicorn>=22.0

# Async tasks
celery>=5.3
redis>=5.0

# PDF generation
weasyprint>=60.0

# HTTP (WhatsApp API)
requests>=2.31

# Image handling
Pillow>=10.0

# Dev tools
django-extensions>=3.2
